backend/requirements.txt
flask==2.0.1
tensorflow==2.9.1
opencv-python==4.6.0.66
transformers==4.21.0
pymongo==4.3.3
python-dotenv==0.19.2
flask-cors==3.0.10
numpy==1.23.5